server/pyproject.toml
[project]
name = "lumen-server"
version = "0.1.0"
description = "Speech-to-3D point cloud generation server"
requires-python = ">=3.14"
dependencies = [
    # Web framework
    "fastapi>=0.129",
    "uvicorn[standard]>=0.41",

    # Data validation + config
    "pydantic>=2.12",
    "pydantic-settings>=2.13",

    # ML runtime
    "torch>=2.10",

    # Numeric / geometry
    "numpy>=2.4",
    "trimesh>=4.11",
    "scipy>=1.17",

    # Image
    "Pillow>=12.1",

    # Infrastructure
    "google-cloud-storage>=2.18",
    "cachetools>=7.0",

    # Observability
    "structlog>=25.4",

    # Rate limiting
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    # Core test runner
    "pytest>=9.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "pytest-xdist>=3.5",           # Parallel test execution (pytest -n auto)
    "httpx>=0.28",

    # Pydantic-aware test data generation
    "polyfactory>=3.2",            # Auto-generate valid instances from Pydantic models

    # Declarative assertions + snapshot testing
    "dirty-equals>=0.11",          # e.g. assert x == IsPositiveInt
    "inline-snapshot>=0.23",       # Inline snapshot() values in source code

    # HTTP mocking (transport-layer, no Docker)
    "respx>=0.22",                 # Mock httpx requests in-process

    # Property-based testing
    "hypothesis>=6.130",           # Invariant testing for encoders, samplers, cache keys

    # Linting + type checking
    "ruff>=0.15",
    "mypy>=1.19",
]
ml = [
    "diffusers>=0.36",
    "transformers>=5.2",
    "accelerate>=1.12",
    "safetensors>=0.5",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"
